# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: APExplorer — API Explorer & Binary Analyzer
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: collections
    description: Request collections
  - name: requests
    description: HTTP requests
  - name: files
    description: File analysis
  - name: tools
    description: Import/parse tools
  - name: dashboard
    description: Dashboard stats
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /collections:
    get:
      operationId: listCollections
      tags: [collections]
      summary: List all collections
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
    post:
      operationId: createCollection
      tags: [collections]
      summary: Create a collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"

  /collections/{id}:
    get:
      operationId: getCollection
      tags: [collections]
      summary: Get a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          description: Not found
    patch:
      operationId: updateCollection
      tags: [collections]
      summary: Update a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
    delete:
      operationId: deleteCollection
      tags: [collections]
      summary: Delete a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List requests, optionally by collection
      parameters:
        - name: collectionId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedRequest"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Create a saved request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedRequest"

  /requests/recent:
    get:
      operationId: listRecentRequests
      tags: [requests]
      summary: Get recently executed requests
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedRequest"

  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedRequest"
        "404":
          description: Not found
    patch:
      operationId: updateRequest
      tags: [requests]
      summary: Update a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedRequest"
    delete:
      operationId: deleteRequest
      tags: [requests]
      summary: Delete a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /requests/{id}/send:
    post:
      operationId: sendRequest
      tags: [requests]
      summary: Execute a saved request and store the response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Response from target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestResult"

  /requests/{id}/history:
    get:
      operationId: getRequestHistory
      tags: [requests]
      summary: Get response history for a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequestResult"

  /tools/parse-curl:
    post:
      operationId: parseCurl
      tags: [tools]
      summary: Parse a cURL command into a request object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CurlInput"
      responses:
        "200":
          description: Parsed request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParsedRequest"

  /tools/send-adhoc:
    post:
      operationId: sendAdhoc
      tags: [tools]
      summary: Send an ad-hoc request without saving
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdhocRequestInput"
      responses:
        "200":
          description: Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestResult"

  /files:
    get:
      operationId: listFileAnalyses
      tags: [files]
      summary: List file analyses
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileAnalysis"

  /files/{id}:
    get:
      operationId: getFileAnalysis
      tags: [files]
      summary: Get a file analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileAnalysis"
        "404":
          description: Not found
    delete:
      operationId: deleteFileAnalysis
      tags: [files]
      summary: Delete a file analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Dashboard overview stats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Collection:
      type: object
      required: [id, name, createdAt, requestCount]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        createdAt:
          type: string
        requestCount:
          type: integer

    CollectionInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string

    CollectionUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string

    SavedRequest:
      type: object
      required: [id, name, method, url, createdAt, updatedAt]
      properties:
        id:
          type: integer
        collectionId:
          type: ["integer", "null"]
        name:
          type: string
        method:
          type: string
        url:
          type: string
        headers:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        queryParams:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        lastStatusCode:
          type: ["integer", "null"]
        lastResponseTime:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    SavedRequestInput:
      type: object
      required: [name, method, url]
      properties:
        collectionId:
          type: ["integer", "null"]
        name:
          type: string
          minLength: 1
        method:
          type: string
        url:
          type: string
        headers:
          type: string
        body:
          type: string
        queryParams:
          type: string
        notes:
          type: string

    SavedRequestUpdate:
      type: object
      properties:
        collectionId:
          type: ["integer", "null"]
        name:
          type: string
        method:
          type: string
        url:
          type: string
        headers:
          type: string
        body:
          type: string
        queryParams:
          type: string
        notes:
          type: string

    RequestResult:
      type: object
      required: [id, requestId, statusCode, statusText, responseHeaders, responseBody, responseTime, executedAt]
      properties:
        id:
          type: integer
        requestId:
          type: integer
        statusCode:
          type: integer
        statusText:
          type: string
        responseHeaders:
          type: string
        responseBody:
          type: string
        responseTime:
          type: integer
        executedAt:
          type: string

    CurlInput:
      type: object
      required: [curl]
      properties:
        curl:
          type: string

    ParsedRequest:
      type: object
      required: [method, url]
      properties:
        method:
          type: string
        url:
          type: string
        headers:
          type: string
        body:
          type: string
        queryParams:
          type: string

    AdhocRequestInput:
      type: object
      required: [method, url]
      properties:
        method:
          type: string
        url:
          type: string
        headers:
          type: string
        body:
          type: string
        queryParams:
          type: string

    FileAnalysis:
      type: object
      required: [id, filename, filesize, mimetype, md5, sha256, createdAt]
      properties:
        id:
          type: integer
        filename:
          type: string
        filesize:
          type: integer
        mimetype:
          type: string
        md5:
          type: string
        sha256:
          type: string
        entropy:
          type: ["number", "null"]
        strings:
          type: ["string", "null"]
        hexDump:
          type: ["string", "null"]
        metadata:
          type: ["string", "null"]
        isSuspicious:
          type: boolean
        suspiciousReasons:
          type: ["string", "null"]
        createdAt:
          type: string

    DashboardStats:
      type: object
      required: [totalCollections, totalRequests, totalFileAnalyses, recentRequestCount, methodBreakdown, statusBreakdown]
      properties:
        totalCollections:
          type: integer
        totalRequests:
          type: integer
        totalFileAnalyses:
          type: integer
        recentRequestCount:
          type: integer
        methodBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/MethodCount"
        statusBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"

    MethodCount:
      type: object
      required: [method, count]
      properties:
        method:
          type: string
        count:
          type: integer

    StatusCount:
      type: object
      required: [range, count]
      properties:
        range:
          type: string
        count:
          type: integer