# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * APExplorer — API Explorer & Binary Analyzer
 * OpenAPI spec version: 0.1.0
 */

export interface Collection {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  createdAt: string;
  requestCount: number;
}