# lib/api-zod/src/generated/types/collectionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * APExplorer — API Explorer & Binary Analyzer
 * OpenAPI spec version: 0.1.0
 */

export interface CollectionInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  color?: string;
}