# lib/api-zod/src/generated/types/curlInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * APExplorer — API Explorer & Binary Analyzer
 * OpenAPI spec version: 0.1.0
 */

export interface CurlInput {
  curl: string;
}